+++ packages/client/src/utils/schema.js
@@ -2,7 +2,7 @@ import { API } from "api"
 import TableFetch from "@budibase/frontend-core/src/fetch/TableFetch"
 import ViewFetch from "@budibase/frontend-core/src/fetch/ViewFetch"
 import QueryFetch from "@budibase/frontend-core/src/fetch/QueryFetch.js"
-import RelationshipFetch from "@budibase/frontend-core/src/fetch/RelationshipFetch.js"
+import RelationshipFetch from "@budibase/frontend-core/src/fetch/RelationshipFetch"
 import NestedProviderFetch from "@budibase/frontend-core/src/fetch/NestedProviderFetch.js"
 import FieldFetch from "@budibase/frontend-core/src/fetch/FieldFetch"
 import JSONArrayFetch from "@budibase/frontend-core/src/fetch/JSONArrayFetch.js"
