+++ packages/frontend-core/src/fetch/index.ts
@@ -2,7 +2,7 @@ import TableFetch from "./TableFetch.js"
 import ViewFetch from "./ViewFetch.js"
 import ViewV2Fetch from "./ViewV2Fetch.js"
 import QueryFetch from "./QueryFetch.js"
-import RelationshipFetch from "./RelationshipFetch.js"
+import RelationshipFetch from "./RelationshipFetch"
 import NestedProviderFetch from "./NestedProviderFetch.js"
 import FieldFetch from "./FieldFetch"
 import JSONArrayFetch from "./JSONArrayFetch.js"
